pyproject.toml
[tool.poetry]
name = "python-otel-experiment"
version = "0.1.0"
description = "Python OTEL Experiment"
authors = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
maintainers = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/face0b1101/python-otel-experiment"
repository = "https://github.com/face0b1101/python-otel-experiment"
documentation = "https://github.com/face0b1101/python-otel-experiment"

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.6"
Unipath = "^1.1"
rich = "^12.5.1"
Flask = "^2.2.2"
opentelemetry-distro = "^0.35b0"
opentelemetry-exporter-otlp = "^1.14.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.14.0"

[tool.poetry.group.dev.dependencies]
poetry-lock-package = "^0.4.5"
pytest = "^7.2.0"
tox = "^3.27.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
# make it compatible with black
profile = "black" 

[tool.poetry.scripts]
python-otel-experiment = 'python_otel_experiment.main:main'
